pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hs-tasnet"
version = "0.1.0"
description = "HS-TasNet training, inference, and export tooling."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Sahal Sajeer Kalandan" }]

dependencies = [
  "torch>=2.1",
  "numpy>=1.23",
  "tqdm>=4.66",
  "pyyaml>=6.0",
  "soundfile>=0.12",
  "tensorboard>=2.15",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.4",
]

[project.scripts]
hs-tasnet = "hs_tasnet.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
